requirements.txt
absl-py==1.0.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.28.1
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
flatbuffers==2.0
fonttools==4.31.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.11.0
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.2.1
jupyter-core==4.9.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
protobuf==3.20.0
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pywin32==303;platform_system == "Windows"
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.1
whitenoise==6.0.0
wrapt==1.14.0
zipp==3.7.0
